static: package static files in the final build
This changes how templates and static files are
usedi by generate all files in blob to build a
binary without dependences.

This simplify the usage by providing just a
single final binary.

Presents two new flags in case you wnat to run
and manage the templates and files localy:
`--templates` set the path to the templates directory
`--static` set the path to the static files directory